Organize files
 - implements String.length simplifier
