+++ ecmascript/transforms/src/compat/es2015/shorthand_property.rs
@@ -1,4 +1,4 @@
-use swc_common::{pos::Mark, Fold, FoldWith};
+use swc_common::{Fold, FoldWith};
 use swc_ecma_ast::*;
 
 /// Compile ES2015 shorthand properties to ES5
@@ -44,7 +44,7 @@ impl Fold<Prop> for Shorthand {
 
         match prop {
             Prop::Shorthand(Ident { sym, span }) => {
-                let span = span.apply_mark(Mark::fresh(Mark::root()));
+                let span = mark!(span);
 
                 Prop::KeyValue(KeyValueProp {
                     key: PropName::Ident(Ident {
