+++ ecmascript/transforms/src/quote.rs
@@ -5,6 +5,16 @@ macro_rules! quote_ident {
     }};
 }
 
+/// Mark span as `processed` to prevent the emitter from deoptimizing
+#[macro_export]
+macro_rules! mark {
+    ($span:expr) => {{
+        let mut span = $span;
+        let parent = span.remove_mark();
+        $span.apply_mark(::swc_common::pos::Mark::fresh(parent))
+    }};
+}
+
 #[macro_export]
 macro_rules! expr {
     ($span:expr, null) => {{
